remote-index-build-service/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
opentelemetry-sdk==1.30.0
boto3==1.36.23